+++ src/remaining/scanner.cc
@@ -1414,7 +1414,7 @@ YY_RULE_SETUP
                                  yylloc.first_line = yylineno;
                                  yylloc.first_column = column;
                                  column += yyleng;
-                                 yylval.rval = stof(yytext);
+                                 yylval.rval = atof(yytext);
                                  return T_REALNUM;
                                  
                                  }
